+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - [helpers] add some lodash-esque functions
 - [helpers] fix types after upgrading typescript
+- [form] reset `isSubmitting` when encountering formErrors
 - [all] remove lodash to decrease bundle size
 - [all] upgrade all dev deps
 - [all] export all interfaces and types